+++ R/ResamplingBootstrap.R
@@ -24,8 +24,6 @@
 #' See [Resampling].
 #'
 #' @section Parameters:
-#' * `stratify` :: `logical(1)` | `character()`\cr
-#'   Enables stratification. See [Resampling].
 #' * `repeats` :: `integer(1)`\cr
 #'   Number of repetitions.
 #' * `ratio` :: `numeric(1)`\cr
@@ -53,7 +51,6 @@ ResamplingBootstrap = R6Class("ResamplingBootstrap", inherit = Resampling,
   public = list(
     initialize = function() {
       ps = ParamSet$new(list(
-        ParamUty$new("stratify", default = NULL),
         ParamInt$new("repeats", lower = 1L, tags = "required"),
         ParamDbl$new("ratio", lower = 0, upper = 1, tags = "required"))
       )