+++ man-roxygen/seealso_learner.R
@@ -1,2 +1,4 @@
 #' @seealso
 #' [Dictionary][mlr3misc::Dictionary] of [Learners][Learner]: [mlr_learners]
+#'
+#' `as.data.table(mlr_learners)` for a complete table of all (also dynamically created) [Learner] implementations.