+++ man-roxygen/seealso_measure.R
@@ -1,2 +1,4 @@
 #' @seealso
 #' [Dictionary][mlr3misc::Dictionary] of [Measures][Measure]: [mlr_measures]
+#'
+#' `as.data.table(mlr_measures)` for a complete table of all (also dynamically created) [Measure] implementations.