+++ man-roxygen/seealso_resampling.R
@@ -1,2 +1,4 @@
 #' @seealso
 #' [Dictionary][mlr3misc::Dictionary] of [Resamplings][Resampling]: [mlr_resamplings]
+#'
+#' `as.data.table(mlr_resamplings)` for a complete table of all (also dynamically created) [Resampling] implementations.