+++ man-roxygen/seealso_task.R
@@ -1,3 +1,4 @@
 #' @seealso
 #' [Dictionary][mlr3misc::Dictionary] of [Tasks][Task]: [mlr_tasks]
-
+#'
+#' `as.data.table(mlr_tasks)` for a complete table of all (also dynamically created) [Tasks][Task].