+++ man-roxygen/seealso_task_generator.R
@@ -1,3 +1,4 @@
 #' @seealso
 #' [Dictionary][mlr3misc::Dictionary] of [TaskGenerators][TaskGenerator]: [mlr_task_generators]
-
+#'
+#' `as.data.table(mlr_resamplings)` for a complete table of all (also dynamically created) [TaskGenerator] implementations.